+++ src/lib/Toaster.svelte
@@ -125,6 +125,7 @@
 		toastOptions = {},
 		dir = 'auto',
 		gap = GAP,
+		pauseWhenPageIsHidden = false,
 		loadingIcon: loadingIconProp,
 		successIcon: successIconProp,
 		errorIcon: errorIconProp,
@@ -278,6 +279,7 @@
 			);
 
 			const changeHandler = ({ matches }: MediaQueryListEvent) => {
+				if (theme !== 'system') return;
 				actualTheme = matches ? DARK : LIGHT;
 			};
 
@@ -308,7 +310,7 @@
 		onfocus?.(event);
 		const isNotDismissable =
 			event.target instanceof HTMLElement &&
-			event.target.dataset.dismissable === 'false';
+			event.target.dataset.dismissible === 'false';
 
 		if (isNotDismissable) return;
 
@@ -324,7 +326,7 @@
 		onpointerdown?.(event);
 		const isNotDismissable =
 			event.target instanceof HTMLElement &&
-			event.target.dataset.dismissable === 'false';
+			event.target.dataset.dismissible === 'false';
 
 		if (isNotDismissable) return;
 		interacting = true;
@@ -436,6 +438,7 @@
 							closeButtonAriaLabel}
 						expandByDefault={expand}
 						{expanded}
+						{pauseWhenPageIsHidden}
 						loadingIcon={loadingIconProp}
 					>
 						{#snippet successIcon()}